+++ resources/region-field/js/components/DetailField.vue
@@ -1,0 +1,18 @@
+<template>
+    <panel-item :field="field">
+        <template slot="value">
+            {{ value }}
+        </template>
+    </panel-item>
+</template>
+
+<script>
+export default {
+  props: ['resource', 'resourceName', 'resourceId', 'field'],
+  computed: {
+    value() {
+      return this.field.value;
+    },
+  },
+};
+</script>