+++ resources/region-field/js/components/IndexField.vue
@@ -1,0 +1,14 @@
+<template>
+  <span>{{ value }}</span>
+</template>
+
+<script>
+export default {
+  props: ['resourceName', 'field'],
+  computed: {
+    value() {
+      return this.field.value;
+    },
+  },
+};
+</script>