+++ resources/region-field/js/region-field.js
@@ -1,0 +1,5 @@
+Nova.booting((Vue, router, store) => {
+  Vue.component('index-region-field', require('./components/IndexField'));
+  Vue.component('detail-region-field', require('./components/DetailField'));
+  Vue.component('form-region-field', require('./components/FormField'));
+});